+++ src/game/actors/attributes/StrengthAttribute.java
@@ -1,7 +1,7 @@
 package actors.attributes;
 
 public class StrengthAttribute extends Attribute {
-    public StrengthAttribute(int value) {
+    public StrengthAttribute(double value) {
         super(AttributeTypes.STRENGTH, value);
     }
 }
