+++ src/game/actors/resistances/DarknessResistance.java
@@ -3,7 +3,7 @@ package actors.resistances;
 import abilities.damages.DamageTypes;
 
 public class DarknessResistance extends Resistance {
-    public DarknessResistance(int value) {
+    public DarknessResistance(double value) {
         super(DamageTypes.DARKNESS, value);
     }
 }
