+++ src/game/actors/resistances/FireResistance.java
@@ -3,7 +3,7 @@ package actors.resistances;
 import abilities.damages.DamageTypes;
 
 public class FireResistance extends Resistance {
-    public FireResistance(int value) {
+    public FireResistance(double value) {
         super(DamageTypes.FIRE, value);
     }
 }
