+++ src/game/actors/resistances/IceResistance.java
@@ -3,7 +3,7 @@ package actors.resistances;
 import abilities.damages.DamageTypes;
 
 public class IceResistance extends Resistance {
-    public IceResistance(int value) {
+    public IceResistance(double value) {
         super(DamageTypes.ICE, value);
     }
 }
