+++ src/game/actors/resistances/LightResistance.java
@@ -3,7 +3,7 @@ package actors.resistances;
 import abilities.damages.DamageTypes;
 
 public class LightResistance extends Resistance {
-    public LightResistance(int value) {
+    public LightResistance(double value) {
         super(DamageTypes.LIGHT, value);
     }
 }
