+++ src/game/actors/resistances/WaterResistance.java
@@ -3,7 +3,7 @@ package actors.resistances;
 import abilities.damages.DamageTypes;
 
 public class WaterResistance extends Resistance {
-    public WaterResistance(int value) {
+    public WaterResistance(double value) {
         super(DamageTypes.WATER, value);
     }
 }
