+++ src/game/actors/resources/StressValues.java
@@ -1,7 +1,7 @@
 package actors.resources;
 
 public class StressValues extends ResourceValues {
-    public StressValues(int maxValue, int value, int regenValue) {
+    public StressValues(double maxValue, double value, double regenValue) {
         super(maxValue, value, regenValue);
     }
